requirements.txt
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==1.11.15
django-admin-honeypot==1.1.0
django-ckeditor==5.4.0
django-crontab==0.7.1
django-js-asset==1.1.0
django-recaptcha2==1.0.3
djangorestframework==3.8.2
idna==2.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
sparkpost==1.3.5
sqlparse==0.2.4
urllib3==1.21.1
validators==0.12.2
